+++ erdblick_app/app/app.module.ts
@@ -90,6 +90,11 @@ import {provideAnimationsAsync} from "@angular/platform-browser/animations/async
 import {providePrimeNG} from "primeng/config";
 import {definePreset} from '@primeng/themes';
 import Aura from "@primeng/themes/aura";
+import {ErdblickViewUIComponent} from "./view.ui.component";
+import {ViewService} from "./view.service";
+import {CameraService} from "./camera.service";
+import {MarkerService} from "./marker.service";
+import {ViewStateService} from "./view.state.service";
 
 export const ErdblickTheme = definePreset(Aura, {
     semantic: {
@@ -191,7 +196,8 @@ export const initializeServices = () => {
         TreeTableFilterPatchDirective,
         StatsDialogComponent,
         SourceDataLayerSelectionDialogComponent,
-        LegalInfoDialogComponent
+        LegalInfoDialogComponent,
+        ErdblickViewUIComponent
     ],
     bootstrap: [
         AppComponent
@@ -281,6 +287,10 @@ export const initializeServices = () => {
         EditorService,
         DataSourcesService,
         RightClickMenuService,
+        ViewStateService,
+        CameraService,
+        ViewService,
+        MarkerService,
         provideHttpClient(),
         provideAnimationsAsync(),
         providePrimeNG({
