+++ erdblick_app/app/view.component.ts
@@ -12,7 +12,6 @@ import {
     UrlTemplateImageryProvider,
     Viewer,
     SceneMode,
-    HeightReference,
     Billboard,
     BillboardCollection,
     Rectangle,
@@ -21,77 +20,35 @@ import {
     GeographicProjection
 } from "./cesium";
 import {ParametersService} from "./parameters.service";
-import {AfterViewInit, Component, ElementRef, OnDestroy, ViewChild} from "@angular/core";
+import {AfterViewInit, Component, OnDestroy} from "@angular/core";
 import {MapService} from "./map.service";
 import {DebugWindow, ErdblickDebugApi} from "./debugapi.component";
-import {FeatureSearchService, SearchResultPrimitiveId} from "./feature.search.service";
+import {FeatureSearchService} from "./feature.search.service";
 import {CoordinatesService} from "./coordinates.service";
 import {JumpTargetService} from "./jump.service";
 import {distinctUntilChanged} from "rxjs";
-import {SearchResultPosition} from "./featurefilter.worker";
 import {InspectionService} from "./inspection.service";
 import {KeyboardService} from "./keyboard.service";
 import {coreLib} from "./wasm";
 import {MenuItem} from "primeng/api";
 import {RightClickMenuService} from "./rightclickmenu.service";
 import {AppModeService} from "./app-mode.service";
+import {ViewService} from "./view.service";
+import {CameraService} from "./camera.service";
+import {MarkerService} from "./marker.service";
+import {ViewStateService} from "./view.state.service";
 
 // Redeclare window with extended interface
 declare let window: DebugWindow;
-
-interface MarkersParams {
-    id?: SearchResultPrimitiveId;
-    position: Cartesian3;
-    image?: string;
-    width: number;
-    height: number;
-    eyeOffset?: Cartesian3;
-    pixelOffset?: Cartesian2;
-    color?: Color;
-    disableDepthTestDistance?: number;
-    heightReference?: HeightReference;
-}
 
 @Component({
     selector: 'erdblick-view',
     template: `
         <div #viewer id="mapViewContainer" class="mapviewer-renderlayer" style="z-index: 0"></div>
-        <div class="navigation-controls" *ngIf="!appModeService.isVisualizationOnly">
-            <div class="nav-control-group">
-                <p-button icon="pi pi-plus" (onClick)="zoomIn()" [rounded]="true" severity="secondary" size="small" pTooltip="Zoom In (Q)"></p-button>
-                <p-button icon="pi pi-minus" (onClick)="zoomOut()" [rounded]="true" severity="secondary" size="small" pTooltip="Zoom Out (E)"></p-button>
-            </div>
-            <div class="nav-control-group">
-                <p-button icon="pi pi-arrow-up" (onClick)="moveUp()" [rounded]="true" severity="secondary" size="small" pTooltip="Move Up (W)"></p-button>
-                <div class="nav-horizontal">
-                    <p-button icon="pi pi-arrow-left" (onClick)="moveLeft()" [rounded]="true" severity="secondary" size="small" pTooltip="Move Left (A)"></p-button>
-                    <p-button icon="pi pi-arrow-right" (onClick)="moveRight()" [rounded]="true" severity="secondary" size="small" pTooltip="Move Right (D)"></p-button>
-                </div>
-                <p-button icon="pi pi-arrow-down" (onClick)="moveDown()" [rounded]="true" severity="secondary" size="small" pTooltip="Move Down (S)"></p-button>
-            </div>
-            <p-button icon="pi pi-refresh" (onClick)="resetOrientation()" [rounded]="true" severity="secondary" size="small" pTooltip="Reset View (R)"></p-button>
-        </div>
-        <div class="compass-circle">
-            <div class="compass-label north">N</div>
-            <div class="compass-label east">E</div>
-            <div class="compass-label south">S</div>
-            <div class="compass-label west">W</div>
-            <div class="compass-needle" #compassNeedle></div>
-        </div>
-        <div class="scene-mode-toggle" *ngIf="!appModeService.isVisualizationOnly">
-            <p-button
-                    [ngClass]="{'blue': is2DMode}"
-                    [label]="is2DMode ? '2D' : '3D'"
-                    [pTooltip]="is2DMode ? 'Switch to 3D' : 'Switch to 2D'"
-                    tooltipPosition="left"
-                    (onClick)="toggleSceneMode()"
-                    [rounded]="true"
-                    severity="secondary"
-                    size="large">
-            </p-button>
-        </div>
-        <p-contextMenu *ngIf="!appModeService.isVisualizationOnly" [target]="viewer" [model]="menuItems" (onHide)="onContextMenuHide()" />
+        <p-contextMenu *ngIf="!appModeService.isVisualizationOnly" [target]="viewer" [model]="menuItems"
+                       (onHide)="onContextMenuHide()"/>
         <sourcedatadialog *ngIf="!appModeService.isVisualizationOnly"></sourcedatadialog>
+        <erdblick-view-ui></erdblick-view-ui>
     `,
     styles: [`
         @media only screen and (max-width: 56em) {
@@ -100,57 +57,14 @@ interface MarkersParams {
                 padding-bottom: 0;
             }
         }
-        .scene-mode-toggle {
-            position: absolute;
-            bottom: 0.5em;
-            right: 1em;
-            z-index: 110;
-        }
-        .navigation-controls {
-            position: absolute;
-            bottom: 4.5em;
-            right: 0.5em;
-            z-index: 1;
-            display: flex;
-            flex-direction: column;
-            gap: 0.5em;
-            align-items: center;
-        }
-        .nav-control-group {
-            display: flex;
-            flex-direction: column;
-            gap: 0.25em;
-            align-items: center;
-        }
-        .nav-horizontal {
-            display: flex;
-            gap: 0.25em;
-        }
     `],
     standalone: false
 })
 export class ErdblickViewComponent implements AfterViewInit, OnDestroy {
-    @ViewChild('compassNeedle', { static: true }) needleRef!: ElementRef<HTMLElement>;
-
-    viewer!: Viewer;
     private mouseHandler: ScreenSpaceEventHandler | null = null;
     private openStreetMapLayer: ImageryLayer | null = null;
-    private markerCollection: BillboardCollection | null = null;
     private tileOutlineEntity: Entity | null = null;
     menuItems: MenuItem[] = [];
-    private cameraIsMoving: boolean = false;
-    private ignoreNextCameraUpdate: boolean = false;
-    is2DMode: boolean;
-    
-    // State to preserve during viewer reinitialization
-    private viewerState: {
-        openStreetMapLayerAlpha: number;
-        openStreetMapLayerShow: boolean;
-        markerPositions: Cartesian3[];
-        tileOutlineEntity: Entity | null;
-        cameraState: any;
-        menuItems: MenuItem[];
-    } | null = null;
 
     /**
      * Construct a Cesium View with a Model.
@@ -165,29 +79,30 @@ export class ErdblickViewComponent implements AfterViewInit, OnDestroy {
      * @param appModeService
      */
     constructor(private mapService: MapService,
-                private featureSearchService: FeatureSearchService,
                 private parameterService: ParametersService,
                 private jumpService: JumpTargetService,
                 private inspectionService: InspectionService,
                 private keyboardService: KeyboardService,
                 private menuService: RightClickMenuService,
                 private coordinatesService: CoordinatesService,
+                private featureSearchService: FeatureSearchService,
+                private viewStateService: ViewStateService,
+                private viewService: ViewService,
+                private cameraService: CameraService,
+                private markerService: MarkerService,
                 public appModeService: AppModeService) {
-
-        this.is2DMode = this.parameterService.parameters.getValue().mode2d;
-
         this.mapService.tileVisualizationTopic.subscribe((tileVis: TileVisualization) => {
             // Safety check: ensure viewer exists and is not destroyed
-            if (!this.viewer || (typeof this.viewer.isDestroyed === 'function' && this.viewer.isDestroyed())) {
-                console.warn('Cannot render tile visualization: viewer not available');
+            if (this.viewStateService.isUnavailable() || this.viewStateService.isDestroyed()) {
+                console.debug('Cannot render tile visualization: viewer not available');
                 return;
             }
-            
-            tileVis.render(this.viewer).then(wasRendered => {
-                if (wasRendered && this.viewer && this.viewer.scene) {
+
+            tileVis.render(this.viewStateService.viewer).then(wasRendered => {
+                if (wasRendered && this.viewStateService.viewer && this.viewStateService.viewer.scene) {
                     // Double-check viewer is still available after async operation
-                    if (typeof this.viewer.isDestroyed === 'function' && !this.viewer.isDestroyed()) {
-                    this.viewer.scene.requestRender();
+                    if (this.viewStateService.isDestroyed()) {
+                        this.viewStateService.viewer.scene.requestRender();
                     }
                 }
             });
@@ -195,79 +110,83 @@ export class ErdblickViewComponent implements AfterViewInit, OnDestroy {
 
         this.mapService.tileVisualizationDestructionTopic.subscribe((tileVis: TileVisualization) => {
             // Safety check: ensure viewer exists and is not destroyed
-            if (!this.viewer || (typeof this.viewer.isDestroyed === 'function' && this.viewer.isDestroyed())) {
-                console.warn('Cannot destroy tile visualization: viewer not available');
+            if (this.viewStateService.isUnavailable() || this.viewStateService.isDestroyed()) {
+                console.debug('Cannot destroy tile visualization: viewer not available');
                 return;
             }
-            
-            tileVis.destroy(this.viewer);
-            if (this.viewer && this.viewer.scene) {
-            this.viewer.scene.requestRender();
+
+            tileVis.destroy(this.viewStateService.viewer);
+            if (this.viewStateService.viewer && this.viewStateService.viewer.scene) {
+                this.viewStateService.viewer.scene.requestRender();
             }
         });
 
-        this.mapService.moveToWgs84PositionTopic.subscribe((pos: {x: number, y: number, z?: number}) => {
+        this.mapService.moveToWgs84PositionTopic.subscribe((pos: { x: number, y: number, z?: number }) => {
             // Safety check: ensure viewer exists and is not destroyed
-            if (!this.viewer || !this.viewer.camera || (typeof this.viewer.isDestroyed === 'function' && this.viewer.isDestroyed())) {
-                console.warn('Cannot move to WGS84 position: viewer not available');
+            if (this.viewStateService.isUnavailable() || this.viewStateService.isDestroyed()) {
+                console.debug('Cannot move to WGS84 position: viewer not available');
                 return;
             }
-            
-            if (this.is2DMode) {
+
+            if (this.viewStateService.is2DMode) {
                 // In 2D mode, create a Rectangle centered on the target position
                 // Use current view rectangle to preserve the exact zoom level
-                const canvas = this.viewer.scene.canvas;
-                let currentRect = this.viewer.camera.computeViewRectangle(this.viewer.scene.globe.ellipsoid);
-                
+                const canvas = this.viewStateService.viewer.scene.canvas;
+                let currentRect = this.viewStateService.viewer.camera.computeViewRectangle(
+                    this.viewStateService.viewer.scene.globe.ellipsoid
+                );
+
                 // If computeViewRectangle fails, use robust calculation
                 if (!currentRect) {
-                    currentRect = this.computeRobustViewRectangle(canvas);
+                    currentRect = this.cameraService.computeRobustViewRectangle(canvas);
                 }
-                
+
                 if (currentRect) {
                     // Calculate the current view size
                     const currentWidth = currentRect.east - currentRect.west;
                     const currentHeight = currentRect.north - currentRect.south;
-                    
+
                     // Center the rectangle on the target position with same dimensions
                     const centerLon = CesiumMath.toRadians(pos.x);
                     const centerLat = CesiumMath.toRadians(pos.y);
                     const halfWidth = currentWidth / 2;
                     const halfHeight = currentHeight / 2;
-                    
+
                     const rectangle = new Rectangle(
                         centerLon - halfWidth,
                         centerLat - halfHeight,
                         centerLon + halfWidth,
                         centerLat + halfHeight
                     );
-                    
+
                     // Ignore the camera change event to preserve mode switch cache
-                    this.ignoreNextCameraUpdate = true;
-                    this.viewer.camera.setView({
+                    this.cameraService.ignoreNextCameraUpdate = true;
+                    this.viewStateService.viewer.camera.setView({
                         destination: rectangle
                     });
                 } else {
                     // Fallback: use position-only movement without changing zoom
-                    const cameraHeight = this.viewer.camera.positionCartographic.height;
-                    this.ignoreNextCameraUpdate = true;
-                    this.viewer.camera.setView({
+                    const cameraHeight = this.viewStateService.viewer.camera.positionCartographic.height;
+                    this.cameraService.ignoreNextCameraUpdate = true;
+                    this.viewStateService.viewer.camera.setView({
                         destination: Cartesian3.fromDegrees(pos.x, pos.y, cameraHeight)
                     });
                 }
             } else {
                 // 3D mode - use current implementation
-            this.parameterService.setView(
-                Cartesian3.fromDegrees(
-                    pos.x,
-                    pos.y,
-                    pos.z !== undefined? pos.z : Cartographic.fromCartesian(this.viewer.camera.position).height),
-                {
-                    heading: CesiumMath.toRadians(0), // East, in radians.
-                    pitch: CesiumMath.toRadians(-90), // Directly looking down.
-                    roll: 0 // No rotation.
-                }
-            );
+                this.parameterService.setView(
+                    Cartesian3.fromDegrees(
+                        pos.x,
+                        pos.y,
+                        pos.z !== undefined ? pos.z : Cartographic.fromCartesian(
+                            this.viewStateService.viewer.camera.position
+                        ).height),
+                    {
+                        heading: CesiumMath.toRadians(0), // East, in radians.
+                        pitch: CesiumMath.toRadians(-90), // Directly looking down.
+                        roll: 0 // No rotation.
+                    }
+                );
             }
         });
 
@@ -278,8 +197,8 @@ export class ErdblickViewComponent implements AfterViewInit, OnDestroy {
 
     ngAfterViewInit() {
         // Initialize viewer with appropriate projection
-        this.createViewer(this.is2DMode).then(() => {
-            // Continue with the rest of the initialization
+        this.createViewer(this.viewStateService.is2DMode).then(() => {
+            this.viewStateService.isViewerInit.next(true);
             this.completeViewerInitialization();
         }).catch((error) => {
             console.error('Failed to initialize viewer:', error);
@@ -292,12 +211,11 @@ export class ErdblickViewComponent implements AfterViewInit, OnDestroy {
      * Complete the viewer initialization process after the viewer is created
      */
     private completeViewerInitialization() {
-        // Setup parameter subscriptions and event handlers
         this.setupParameterSubscriptions();
         this.setupEventHandlers();
         this.setupAdditionalSubscriptions();
         this.setupKeyboardShortcuts();
-        
+
         // Hide the global loading spinner
         const spinner = document.getElementById('global-spinner-container');
         if (spinner) {
@@ -310,18 +228,18 @@ export class ErdblickViewComponent implements AfterViewInit, OnDestroy {
      */
     private setupParameterSubscriptions() {
         this.parameterService.cameraViewData.pipe(distinctUntilChanged()).subscribe(cameraData => {
-            this.ignoreNextCameraUpdate = true;
-            if (this.is2DMode) {
+            this.cameraService.ignoreNextCameraUpdate = true;
+            if (this.viewStateService.is2DMode) {
                 // In 2D mode, check if we have a view rectangle in parameters
                 const params = this.parameterService.p();
                 if (params.viewRectangle && params.viewRectangle.length === 4) {
-                    this.viewer.camera.setView({
+                    this.viewStateService.viewer.camera.setView({
                         destination: Rectangle.fromDegrees(...params.viewRectangle)
                     });
                 } else {
                     // Fallback to center position
                     const cartographic = Cartographic.fromCartesian(cameraData.destination);
-                    this.viewer.camera.setView({
+                    this.viewStateService.viewer.camera.setView({
                         destination: Rectangle.fromDegrees(
                             CesiumMath.toDegrees(cartographic.longitude) - 1,
                             CesiumMath.toDegrees(cartographic.latitude) - 1,
@@ -332,12 +250,12 @@ export class ErdblickViewComponent implements AfterViewInit, OnDestroy {
                 }
             } else {
                 // 3D mode
-                this.viewer.camera.setView({
+                this.viewStateService.viewer.camera.setView({
                     destination: cameraData.destination,
                     orientation: cameraData.orientation
                 });
             }
-            this.updateViewport();
+            this.viewService.updateViewport();
         });
 
         this.parameterService.parameters.subscribe(parameters => {
@@ -345,13 +263,13 @@ export class ErdblickViewComponent implements AfterViewInit, OnDestroy {
                 this.openStreetMapLayer.show = parameters.osm;
                 this.updateOpenStreetMapLayer(parameters.osmOpacity / 100);
             }
-            if (this.viewer && this.is2DMode !== parameters.mode2d) {
+            if (this.viewStateService.viewer && this.viewStateService.is2DMode !== parameters.mode2d) {
                 // Handle async mode change properly
                 this.applySceneModeChange(parameters.mode2d).catch(error => {
                     console.error('Failed to change scene mode:', error);
                 });
             }
-            
+
             // Handle marker parameters - try immediately, but don't retry here
             // The viewerReinitializationComplete subscription will handle restoration after mode changes
             if (parameters.marker && parameters.markedPosition.length == 2) {
@@ -359,20 +277,11 @@ export class ErdblickViewComponent implements AfterViewInit, OnDestroy {
                     Number(parameters.markedPosition[0]),
                     Number(parameters.markedPosition[1])
                 );
-                
-                this.addMarker(markerPosition);
+
+                this.markerService.addMarker(markerPosition);
             } else {
                 // Clear markers when marker is disabled or no position
-                if (this.markerCollection) {
-                    try {
-                        this.markerCollection.removeAll();
-                        if (this.viewer && this.viewer.scene) {
-                            this.viewer.scene.requestRender();
-                        }
-                    } catch (e) {
-                        console.warn('Error clearing markers:', e);
-                    }
-                }
+                this.markerService.clearMarkers();
             }
         });
     }
@@ -387,9 +296,9 @@ export class ErdblickViewComponent implements AfterViewInit, OnDestroy {
             if (this.appModeService.isVisualizationOnly) return;
 
             const position = movement.position;
-            const cartesian = this.viewer.camera.pickEllipsoid(
+            const cartesian = this.viewStateService.viewer.camera.pickEllipsoid(
                 new Cartesian2(position.x, position.y),
-                this.viewer.scene.globe.ellipsoid
+                this.viewStateService.viewer.scene.globe.ellipsoid
             );
             if (defined(cartesian)) {
                 const cartographic = Cartographic.fromCartesian(cartesian);
@@ -409,7 +318,7 @@ export class ErdblickViewComponent implements AfterViewInit, OnDestroy {
             if (this.appModeService.isVisualizationOnly) return;
 
             const position = movement.position;
-            let feature = this.viewer.scene.pick(position);
+            let feature = this.viewStateService.viewer.scene.pick(position);
             if (defined(feature) && feature.primitive instanceof Billboard && feature.primitive.id.type === "SearchResult") {
                 if (feature.primitive.id) {
                     const featureInfo = this.featureSearchService.searchResults[feature.primitive.id.index];
@@ -441,7 +350,9 @@ export class ErdblickViewComponent implements AfterViewInit, OnDestroy {
             // Handle position update after highlighting, because otherwise
             // there is a race condition between the parameter updates for
             // feature selection and position update.
-            const coordinates = this.viewer.camera.pickEllipsoid(position, this.viewer.scene.globe.ellipsoid);
+            const coordinates = this.viewStateService.viewer.camera.pickEllipsoid(
+                position, this.viewStateService.viewer.scene.globe.ellipsoid
+            );
             if (coordinates !== undefined) {
                 this.coordinatesService.mouseClickCoordinates.next(Cartographic.fromCartesian(coordinates));
             }
@@ -456,33 +367,27 @@ export class ErdblickViewComponent implements AfterViewInit, OnDestroy {
                 return;
             }
             // Do not handle mouse move here if the camera is currently being moved.
-            if (this.cameraIsMoving) {
+            if (this.cameraService.cameraIsMoving) {
                 return;
             }
 
             if (!this.appModeService.isVisualizationOnly) {
-                const coordinates = this.viewer.camera.pickEllipsoid(position, this.viewer.scene.globe.ellipsoid);
+                const coordinates = this.viewStateService.viewer.camera.pickEllipsoid(
+                    position, this.viewStateService.viewer.scene.globe.ellipsoid
+                );
                 if (coordinates !== undefined) {
                     this.coordinatesService.mouseMoveCoordinates.next(Cartographic.fromCartesian(coordinates))
                 }
             }
 
             if (!this.appModeService.isVisualizationOnly) {
-                let feature = this.viewer.scene.pick(position);
+                let feature = this.viewStateService.viewer.scene.pick(position);
                 this.mapService.highlightFeatures(
                     Array.isArray(feature?.id) ? feature.id : [feature?.id],
                     false,
                     coreLib.HighlightMode.HOVER_HIGHLIGHT).then();
             }
         }, ScreenSpaceEventType.MOUSE_MOVE);
-
-        const needle = this.needleRef.nativeElement;
-        this.viewer.clock.onTick.addEventListener(() => {
-            if (needle) {
-                const headingRad = this.viewer.camera.heading;
-                needle.style.transform = `rotate(${headingRad}rad)`;
-            }
-        });
     }
 
     /**
@@ -493,12 +398,12 @@ export class ErdblickViewComponent implements AfterViewInit, OnDestroy {
         window.ebDebug = new ErdblickDebugApi(this.mapService, this.parameterService, this);
 
         this.featureSearchService.visualizationChanged.subscribe(_ => {
-            this.renderFeatureSearchResultTree(this.mapService.zoomLevel.getValue());
-            this.viewer.scene.requestRender();
+            this.markerService.renderFeatureSearchResultTree(this.mapService.zoomLevel.getValue());
+            this.viewStateService.viewer.scene.requestRender();
         });
 
         this.mapService.zoomLevel.pipe(distinctUntilChanged()).subscribe(level => {
-            this.renderFeatureSearchResultTree(level);
+            this.markerService.renderFeatureSearchResultTree(level);
         });
 
         this.jumpService.markedPosition.subscribe(position => {
@@ -514,12 +419,14 @@ export class ErdblickViewComponent implements AfterViewInit, OnDestroy {
             const endPoint = Cartesian3.add(origin, direction, new Cartesian3());
             Cartesian3.normalize(direction, direction);
             Cartesian3.negate(direction, direction);
-            const up = this.viewer.scene.globe.ellipsoid.geodeticSurfaceNormal(endPoint, new Cartesian3());
+            const up = this.viewStateService.viewer.scene.globe.ellipsoid.geodeticSurfaceNormal(
+                endPoint, new Cartesian3()
+            );
             const right = Cartesian3.cross(direction, up, new Cartesian3());
             Cartesian3.normalize(right, right);
             const cameraUp = Cartesian3.cross(right, direction, new Cartesian3());
             Cartesian3.normalize(cameraUp, cameraUp);
-            this.viewer.camera.flyTo({
+            this.viewStateService.viewer.camera.flyTo({
                 destination: endPoint,
                 orientation: {
                     direction: direction,
@@ -530,11 +437,11 @@ export class ErdblickViewComponent implements AfterViewInit, OnDestroy {
 
         this.menuService.tileOutline.subscribe(entity => {
             if (entity) {
-                this.tileOutlineEntity = this.viewer.entities.add(entity);
-                this.viewer.scene.requestRender();
+                this.tileOutlineEntity = this.viewStateService.viewer.entities.add(entity);
+                this.viewStateService.viewer.scene.requestRender();
             } else if (this.tileOutlineEntity) {
-                this.viewer.entities.remove(this.tileOutlineEntity);
-                this.viewer.scene.requestRender();
+                this.viewStateService.viewer.entities.remove(this.tileOutlineEntity);
+                this.viewStateService.viewer.scene.requestRender();
             }
         });
     }
@@ -544,195 +451,13 @@ export class ErdblickViewComponent implements AfterViewInit, OnDestroy {
      */
     private setupKeyboardShortcuts() {
         if (!this.appModeService.isVisualizationOnly) {
-            this.keyboardService.registerShortcut('q', this.zoomIn.bind(this), true);
-            this.keyboardService.registerShortcut('e', this.zoomOut.bind(this), true);
-            this.keyboardService.registerShortcut('w', this.moveUp.bind(this), true);
-            this.keyboardService.registerShortcut('a', this.moveLeft.bind(this), true);
-            this.keyboardService.registerShortcut('s', this.moveDown.bind(this), true);
-            this.keyboardService.registerShortcut('d', this.moveRight.bind(this), true);
-            this.keyboardService.registerShortcut('r', this.resetOrientation.bind(this), true);
-            this.keyboardService.registerShortcut('t', this.toggleSceneMode.bind(this), true);
-        }
-    }
-
-    /**
-     * Update the visible viewport, and communicate it to the model.
-     */
-    updateViewport() {
-        // Safety check for viewer existence
-        if (!this.viewer || !this.viewer.scene || !this.viewer.camera) {
-            console.warn('Cannot update viewport: viewer not available');
-            return;
-        }
-
-        try {
-            // Check if the viewer is destroyed
-            if (typeof this.viewer.isDestroyed === 'function' && this.viewer.isDestroyed()) {
-                console.warn('Cannot update viewport: viewer is destroyed');
-                return;
-            }
-
-        let canvas = this.viewer.scene.canvas;
-        if (!canvas) {
-            console.warn('Cannot update viewport: canvas not available');
-            return;
-        }
-
-        let center = new Cartesian2(canvas.clientWidth / 2, canvas.clientHeight / 2);
-        let centerCartesian = this.viewer.camera.pickEllipsoid(center);
-        let centerLon, centerLat;
-
-        if (centerCartesian !== undefined) {
-            let centerCartographic = Cartographic.fromCartesian(centerCartesian);
-            centerLon = CesiumMath.toDegrees(centerCartographic.longitude);
-            centerLat = CesiumMath.toDegrees(centerCartographic.latitude);
-        } else {
-            let cameraCartographic = Cartographic.fromCartesian(this.viewer.camera.positionWC);
-            centerLon = CesiumMath.toDegrees(cameraCartographic.longitude);
-            centerLat = CesiumMath.toDegrees(cameraCartographic.latitude);
-        }
-
-        // First try: Pass ellipsoid explicitly (workaround for Cesium issue)
-        let rectangle = this.viewer.camera.computeViewRectangle(this.viewer.scene.globe.ellipsoid);
-        
-        if (!rectangle) {
-            // Workaround: Robust rectangle calculation with multiple sample points
-            rectangle = this.computeRobustViewRectangle(canvas);
-        }
-        
-        if (!rectangle) {
-            // Final fallback: Calculate viewport from camera position and height
-            const cameraCartographic = this.viewer.camera.positionCartographic;
-            const cameraLon = CesiumMath.toDegrees(cameraCartographic.longitude);
-            const cameraLat = CesiumMath.toDegrees(cameraCartographic.latitude);
-            const cameraHeight = cameraCartographic.height;
-            
-            // Calculate viewport size based on camera height
-            const earthRadius = this.viewer.scene.globe.ellipsoid.maximumRadius;
-            const visualAngularSize = 2 * Math.atan(cameraHeight / (2 * earthRadius));
-            const visualScale = CesiumMath.toDegrees(visualAngularSize);
-            
-            // Create a reasonable viewport around camera position
-            const halfWidth = visualScale / 2;
-            const halfHeight = visualScale * (canvas.clientHeight / canvas.clientWidth) / 2;
-            
-            rectangle = Rectangle.fromDegrees(
-                cameraLon - halfWidth,
-                cameraLat - halfHeight,
-                cameraLon + halfWidth,
-                cameraLat + halfHeight
-            );
-        }
-        
-        // Clamp to valid WebMercator range (±85.05113°)
-        const maxLat = 85.05113; // WebMercatorProjection.MaximumLatitude
-        rectangle = new Rectangle(
-            rectangle.west,
-            Math.max(rectangle.south, CesiumMath.toRadians(-maxLat)),
-            rectangle.east,
-            Math.min(rectangle.north, CesiumMath.toRadians(maxLat))
-        );
-
-        let west = CesiumMath.toDegrees(rectangle.west);
-        let south = CesiumMath.toDegrees(rectangle.south);
-        let east = CesiumMath.toDegrees(rectangle.east);
-        let north = CesiumMath.toDegrees(rectangle.north);
-        let sizeLon = east - west;
-        let sizeLat = north - south;
-
-        // Check for suspicious viewport dimensions
-        if (Math.abs(sizeLon) > 360 || Math.abs(sizeLat) > 180) {
-            console.error('Suspicious viewport dimensions:', { sizeLon, sizeLat });
-        }
-
-        // For WebMercator 2D mode, use visual scale for proper zoom level calculation
-        if (this.is2DMode) {
-            // Calculate dimensions that represent the visual scale for accurate zoom level detection
-            // In WebMercator at high latitudes, geographic bounds are larger than the visual area
-            const cameraHeight = this.viewer.camera.positionCartographic.height;
-            const earthRadius = this.viewer.scene.globe.ellipsoid.maximumRadius;
-            
-            // Validate camera height is reasonable
-            if (!isFinite(cameraHeight) || cameraHeight <= 0) {
-                console.error('Invalid camera height:', cameraHeight);
-                return;
-            }
-            
-            try {
-                // Derive visual scale from camera height
-                // This represents the visual angular size regardless of projection distortion
-                const visualAngularSize = 2 * Math.atan(cameraHeight / (2 * earthRadius));
-                const visualScale = CesiumMath.toDegrees(visualAngularSize);
-                
-                // Validate visual scale is reasonable
-                if (!isFinite(visualScale) || visualScale <= 0) {
-                    console.error('Invalid visual scale:', visualScale);
-                    return;
-                }
-                
-                // Use visual scale for dimensions to ensure correct zoom level calculation
-                sizeLon = visualScale;
-                sizeLat = visualScale * (canvas.clientHeight / canvas.clientWidth);
-                
-                // Apply reasonable bounds to prevent extreme values
-                sizeLon = Math.max(0.001, Math.min(360, sizeLon));
-                sizeLat = Math.max(0.001, Math.min(180, sizeLat));
-                
-            } catch (error) {
-                console.error('Error in visual scale calculation:', error);
-                // Use geographic bounds as fallback
-                sizeLon = east - west;
-                sizeLat = north - south;
-            }
-        }
-
-        // Don't handle antimeridian - let Cesium's continuous chaining work naturally
-        // The sizeLon can be any value in continuous 2D mode
-
-        // Final validation of all viewport parameters
-        if (!isFinite(centerLon) || !isFinite(centerLat) || 
-            !isFinite(west) || !isFinite(east) || !isFinite(south) || !isFinite(north) ||
-            !isFinite(sizeLon) || !isFinite(sizeLat)) {
-            console.error('Invalid viewport parameters detected, skipping update:', {
-                centerLon, centerLat, west, east, south, north, sizeLon, sizeLat
-            });
-            return;
-        }
-
-        // Ensure dimensions are positive
-        if (sizeLon <= 0 || sizeLat <= 0) {
-            console.error('Invalid viewport dimensions:', { sizeLon, sizeLat });
-            return;
-        }
-
-        // Grow the viewport rectangle by 25%
-        let expandLon = sizeLon * 0.25;
-        let expandLat = sizeLat * 0.25;
-        
-        const viewportData = {
-            south: south - expandLat,
-            west: west - expandLon,
-            width: sizeLon + expandLon * 2,
-            height: sizeLat + expandLat * 2,
-            camPosLon: centerLon,
-            camPosLat: centerLat,
-            orientation: -this.viewer.camera.heading + Math.PI * .5,
-        };
-        
-        // Final validation of viewport data
-        if (!isFinite(viewportData.south) || !isFinite(viewportData.west) ||
-            !isFinite(viewportData.width) || !isFinite(viewportData.height) ||
-            !isFinite(viewportData.camPosLon) || !isFinite(viewportData.camPosLat) ||
-            !isFinite(viewportData.orientation)) {
-            console.error('Invalid viewport data calculated, skipping update:', viewportData);
-            return;
-        }
-        
-        this.mapService.setViewport(viewportData);
-        
-        } catch (error) {
-            console.error('Error updating viewport:', error);
-            console.error('Error stack:', (error as Error)?.stack || 'No stack trace available');
+            this.keyboardService.registerShortcut('q', this.cameraService.zoomIn.bind(this), true);
+            this.keyboardService.registerShortcut('e', this.cameraService.zoomOut.bind(this), true);
+            this.keyboardService.registerShortcut('w', this.cameraService.moveUp.bind(this), true);
+            this.keyboardService.registerShortcut('a', this.cameraService.moveLeft.bind(this), true);
+            this.keyboardService.registerShortcut('s', this.cameraService.moveDown.bind(this), true);
+            this.keyboardService.registerShortcut('d', this.cameraService.moveRight.bind(this), true);
+            this.keyboardService.registerShortcut('r', this.cameraService.resetOrientation.bind(this), true);
         }
     }
 
@@ -744,287 +469,9 @@ export class ErdblickViewComponent implements AfterViewInit, OnDestroy {
     }
 
     updateOpenStreetMapLayer(opacity: number) {
-        if (this.openStreetMapLayer && this.viewer && this.viewer.scene) {
+        if (this.openStreetMapLayer && this.viewStateService.viewer && this.viewStateService.viewer.scene) {
             this.openStreetMapLayer.alpha = opacity;
-            this.viewer.scene.requestRender();
-        }
-    }
-
-    addMarker(cartesian: Cartesian3) {
-        // Ensure collection and viewer exist
-        if (!this.markerCollection || !this.viewer || !this.viewer.scene) {
-            console.warn('Cannot add marker: MarkerCollection or viewer not initialized');
-            return false;
-        }
-
-        // Check if the viewer is destroyed
-        if (typeof this.viewer.isDestroyed === 'function' && this.viewer.isDestroyed()) {
-            console.warn('Cannot add marker: viewer is destroyed');
-            return false;
-        }
-        
-        // Clear any existing markers in the collection
-        try {
-        this.markerCollection.removeAll();
-        } catch (e) {
-            console.warn('Error clearing markers:', e);
-            return false;
-        }
-        
-        // Add marker using the same approach as search results
-        try {
-            const params: MarkersParams = {
-                position: cartesian,
-                image: this.featureSearchService.markerGraphics(),
-                width: 32,
-                height: 32
-            };
-            if (this.is2DMode) {
-                params.disableDepthTestDistance = Number.POSITIVE_INFINITY;
-            } else {
-                params.pixelOffset = new Cartesian2(0, -10);
-                params.eyeOffset = new Cartesian3(0, 0, -50);
-                params.heightReference = HeightReference.CLAMP_TO_GROUND;
-            }
-            
-            this.markerCollection.add(params);
-            
-            // Ensure the marker collection is properly added to the scene
-            if (this.viewer.scene.primitives && !this.viewer.scene.primitives.contains(this.markerCollection)) {
-                this.viewer.scene.primitives.add(this.markerCollection);
-            }
-            
-            if (this.viewer.scene.primitives) {
-            this.viewer.scene.primitives.raiseToTop(this.markerCollection);
-            }
-            
-            // Request a render to ensure the marker is visible
-            this.viewer.scene.requestRender();
-            
-            console.debug('Focus marker added successfully');
-            return true;
-            
-        } catch (e) {
-            console.error('Error adding marker:', e);
-            return false;
-        }
-    }
-
-    renderFeatureSearchResultTree(level: number) {
-        if (!this.viewer || !this.viewer.scene) {
-            console.warn('Cannot render feature search results: viewer not initialized');
-            return;
-        }
-
-        // Check if the viewer is destroyed
-        if (typeof this.viewer.isDestroyed === 'function' && this.viewer.isDestroyed()) {
-            console.warn('Cannot render feature search results: viewer is destroyed');
-            return;
-        }
-
-        try {
-        this.featureSearchService.visualization.removeAll();
-        const color = Color.fromCssColorString(this.featureSearchService.pointColor);
-        let markers: Array<[SearchResultPrimitiveId, SearchResultPosition]> = [];
-            
-        // Use the level parameter directly - backend now receives correct viewport coordinates
-        const nodes = this.featureSearchService.resultTree.getNodesAtLevel(level);
-            
-        for (const node of nodes) {
-            if (node.markers.length) {
-                markers.push(...node.markers);
-            } else if (node.count > 0 && node.center) {
-                // For cluster centers, always use the center position directly
-                // The backend coordinates are now correctly aligned with the projection
-                const params: MarkersParams = {
-                    position: node.center,
-                    image: this.featureSearchService.getPinGraphics(node.count),
-                    width: 64,
-                    height: 64
-                };
-                if (this.is2DMode) {
-                    params.disableDepthTestDistance = Number.POSITIVE_INFINITY;
-                } else {
-                    params.eyeOffset = new Cartesian3(0, 0, -50);
-                }
-                this.featureSearchService.visualization.add(params);
-            }
-        }
-
-        if (markers.length) {
-            markers.forEach(marker => {
-                // Always use cartographicRad if available, otherwise fall back to cartesian
-                // This ensures consistent positioning across projections
-                let markerPosition: Cartesian3;
-                if (marker[1].cartographicRad) {
-                    markerPosition = Cartesian3.fromRadians(
-                        marker[1].cartographicRad.longitude,
-                        marker[1].cartographicRad.latitude,
-                        marker[1].cartographicRad.height
-                    );
-                } else {
-                    markerPosition = marker[1].cartesian as Cartesian3;
-                }
-
-                const params: MarkersParams = {
-                    id: marker[0],
-                    position: markerPosition,
-                    image: this.featureSearchService.markerGraphics(),
-                    width: 32,
-                    height: 32,
-                    color: color
-                };
-                if (this.is2DMode) {
-                    params.disableDepthTestDistance = Number.POSITIVE_INFINITY;
-                } else {
-                    params.pixelOffset = new Cartesian2(0, -10);
-                    params.eyeOffset = new Cartesian3(0, 0, -50);
-                }
-                this.featureSearchService.visualization.add(params);
-            });
-        }
-                
-        if (this.viewer && this.viewer.scene && this.viewer.scene.primitives) {
-        this.viewer.scene.primitives.raiseToTop(this.featureSearchService.visualization);
-        }
-        } catch (error) {
-            console.error('Error rendering feature search result tree:', error);
-        }
-    }
-
-    moveUp() {
-        if (this.is2DMode) {
-            const distance = this.get2DMovementDistance();
-            this.moveCameraOnSurface(0, distance.latitudeOffset);
-        } else {
-            this.moveCameraOnSurface(0, this.parameterService.cameraMoveUnits);
-        }
-    }
-
-    moveDown() {
-        if (this.is2DMode) {
-            const distance = this.get2DMovementDistance();
-            this.moveCameraOnSurface(0, -distance.latitudeOffset);
-        } else {
-            this.moveCameraOnSurface(0, -this.parameterService.cameraMoveUnits);
-        }
-    }
-
-    moveLeft() {
-        if (this.is2DMode) {
-            const distance = this.get2DMovementDistance();
-            this.moveCameraOnSurface(-distance.longitudeOffset, 0);
-        } else {
-            this.moveCameraOnSurface(-this.parameterService.cameraMoveUnits, 0);
-        }
-    }
-
-    moveRight() {
-        if (this.is2DMode) {
-            const distance = this.get2DMovementDistance();
-            this.moveCameraOnSurface(distance.longitudeOffset, 0);
-        } else {
-            this.moveCameraOnSurface(this.parameterService.cameraMoveUnits, 0);
-        }
-    }
-
-    private moveCameraOnSurface(longitudeOffset: number, latitudeOffset: number) {
-        // Safety check for viewer existence
-        if (!this.viewer || !this.viewer.camera) {
-            console.warn('Cannot move camera: viewer not available');
-            return;
-        }
-
-        try {
-            // Check if the viewer is destroyed
-            if (typeof this.viewer.isDestroyed === 'function' && this.viewer.isDestroyed()) {
-                console.warn('Cannot move camera: viewer is destroyed');
-                return;
-            }
-
-        // Get the current camera position in Cartographic coordinates (longitude, latitude, height)
-        const cameraPosition = this.viewer.camera.positionCartographic;
-        const lon = cameraPosition.longitude + CesiumMath.toRadians(longitudeOffset);
-        const lat = cameraPosition.latitude + CesiumMath.toRadians(latitudeOffset);
-        const alt = cameraPosition.height;
-        
-        if (this.is2DMode) {
-            // In 2D mode, use setView to maintain the 2D constraints
-            // Ignore the camera change event to preserve mode switch cache
-            this.ignoreNextCameraUpdate = true;
-            this.viewer.camera.setView({
-                destination: Cartesian3.fromRadians(lon, lat, alt)
-            });
-        } else {
-            // 3D mode - use parameter service
-            const newPosition = Cartesian3.fromRadians(lon, lat, alt);
-            this.parameterService.setView(newPosition, this.parameterService.getCameraOrientation());
-            }
-        } catch (error) {
-            console.warn('Error moving camera:', error);
-        }
-    }
-
-    zoomIn() {
-        if (!this.viewer || !this.viewer.camera) {
-            console.warn('Cannot zoom in: viewer not available');
-            return;
-        }
-
-        try {
-        if (this.is2DMode) {
-            this.zoom2D(0.8); // Zoom in by 20%
-        } else {
-            this.viewer.camera.zoomIn(this.parameterService.cameraZoomUnits);
-            }
-        } catch (error) {
-            console.warn('Error zooming in:', error);
-        }
-    }
-
-    zoomOut() {
-        if (!this.viewer || !this.viewer.camera) {
-            console.warn('Cannot zoom out: viewer not available');
-            return;
-        }
-
-        try {
-        if (this.is2DMode) {
-            this.zoom2D(1.25); // Zoom out by 25%
-        } else {
-            this.viewer.camera.zoomOut(this.parameterService.cameraZoomUnits);
-            }
-        } catch (error) {
-            console.warn('Error zooming out:', error);
-        }
-    }
-
-    resetOrientation() {
-        if (!this.viewer || !this.viewer.camera) {
-            console.warn('Cannot reset orientation: viewer not available');
-            return;
-        }
-
-        try {
-        if (this.is2DMode) {
-            // In 2D mode, just reset to north-up orientation
-            // Ignore the camera change event to preserve mode switch cache
-            this.ignoreNextCameraUpdate = true;
-            this.parameterService.setView(this.parameterService.getCameraPosition(), {
-                heading: 0,
-                pitch: CesiumMath.toRadians(-90.0),
-                roll: 0.0
-            });
-        } else {
-            // In 3D mode, reset to default view angle
-            this.parameterService.setView(this.parameterService.getCameraPosition(), {
-                heading: CesiumMath.toRadians(0.0),
-                pitch: CesiumMath.toRadians(-45.0), // 45-degree angle for 3D
-                roll: 0.0
-            });
-            }
-        } catch (error) {
-            console.warn('Error resetting orientation:', error);
+            this.viewStateService.viewer.scene.requestRender();
         }
     }
 
@@ -1034,50 +481,36 @@ export class ErdblickViewComponent implements AfterViewInit, OnDestroy {
         }
     }
 
-    toggleSceneMode() {
-        // Don't allow toggling if already changing modes
-        if (this._isChangingMode) {
-            console.warn('Mode change already in progress, ignoring toggle request');
-            return;
-        }
-        
-        this.parameterService.setCameraMode(!this.is2DMode);
-    }
-
     /**
      * Updated scene mode change to use viewer reinitialization
      */
     private async applySceneModeChange(is2D: boolean) {
         // Prevent multiple mode changes at once
-        if (this._isChangingMode) {
-            console.warn('Mode change already in progress');
+        if (this.viewStateService.isChangingMode) {
+            console.debug('Mode change already in progress');
             return;
         }
 
         // Prevent mode change during destruction
-        if (this._isDestroyingViewer) {
-            console.warn('Mode change prevented: viewer destruction in progress');
+        if (this.viewStateService.isDestroyingViewer) {
+            console.debug('Mode change prevented: viewer destruction in progress');
             return;
         }
 
-        this._isChangingMode = true;
-        
+        this.viewStateService.isChangingMode = true;
+
         try {
             // Recreate viewer with appropriate projection
             await this.recreateViewerForMode(is2D);
             this.setupSceneMode(is2D);
-            this.restoreParameterMarker();
+            this.markerService.restoreParameterMarker();
         } catch (error) {
             console.error('Error during scene mode change:', error);
-            console.warn('Scene mode change failed. Retrying with fallback...');
+            console.debug('Scene mode change failed. Retrying with fallback...');
         } finally {
-            this._isChangingMode = false;
+            this.viewStateService.isChangingMode = false;
         }
     }
-
-    // Add a flag to prevent concurrent mode changes
-    private _isChangingMode = false;
-    private _isDestroyingViewer = false;
 
     /**
      * Recreate the viewer with different projection for 2D/3D modes
@@ -1085,14 +518,14 @@ export class ErdblickViewComponent implements AfterViewInit, OnDestroy {
      */
     private async recreateViewerForMode(is2D: boolean) {
         // Prevent multiple simultaneous reinitializations
-        if (this.viewerState) {
-            console.warn('Viewer reinitialization already in progress');
+        if (this.viewStateService.viewerState) {
+            console.debug('Viewer reinitialization already in progress');
             return;
         }
 
         // Also check if we're currently destroying
-        if (this._isDestroyingViewer) {
-            console.warn('Cannot reinitialize: viewer destruction in progress');
+        if (this.viewStateService.isDestroyingViewer) {
+            console.debug('Cannot reinitialize: viewer destruction in progress');
             return;
         }
 
@@ -1107,12 +540,12 @@ export class ErdblickViewComponent implements AfterViewInit, OnDestroy {
         } catch (error) {
             console.error('Error during viewer reinitialization:', error);
             // Reset state on error to prevent future issues
-            this.viewerState = null;
-            this._isDestroyingViewer = false;
-            
+            this.viewStateService.viewerState = null;
+            this.viewStateService.isDestroyingViewer = false;
+
             // Try to create a basic viewer as fallback
             try {
-                console.log('Attempting fallback viewer creation...');
+                console.warn('Attempting fallback viewer creation...');
                 await this.createViewer(is2D);
             } catch (fallbackError) {
                 console.error('Fallback viewer creation failed:', fallbackError);
@@ -1125,66 +558,32 @@ export class ErdblickViewComponent implements AfterViewInit, OnDestroy {
      * Save the current viewer state before reinitialization
      */
     private saveViewerState() {
-        if (!this.viewer) {
-            console.warn('Cannot save viewer state: viewer is null');
-            return;
-        }
-
         try {
-            if (typeof this.viewer.isDestroyed === 'function' && this.viewer.isDestroyed()) {
-                console.warn('Cannot save viewer state: viewer is destroyed');
+            if (this.viewStateService.isUnavailable() || this.viewStateService.isDestroyed()) {
+                console.debug('Cannot save viewer state: viewer is destroyed or unavailable');
                 return;
             }
 
             const markerPositions: Cartesian3[] = [];
-            if (this.markerCollection) {
-                try {
-                    for (let i = 0; i < this.markerCollection.length; i++) {
-                        const marker = this.markerCollection.get(i);
-                        if (marker && marker.position) {
-                            markerPositions.push(marker.position);
-                        }
+            if (this.markerService.markerCollection) {
+                for (let i = 0; i < this.markerService.markerCollection.length; i++) {
+                    const marker = this.markerService.markerCollection.get(i);
+                    if (marker && marker.position) {
+                        markerPositions.push(marker.position);
                     }
-                } catch (e) {
-                    console.warn('Error collecting marker positions:', e);
                 }
             }
 
-            this.viewerState = {
+            this.viewStateService.viewerState = {
                 openStreetMapLayerAlpha: this.openStreetMapLayer?.alpha || 0.3,
                 openStreetMapLayerShow: this.openStreetMapLayer?.show || false,
                 markerPositions: markerPositions,
                 tileOutlineEntity: this.tileOutlineEntity,
-                cameraState: this.getCurrentCameraState(),
+                cameraState: this.cameraService.getCurrentCameraState(),
                 menuItems: [...this.menuItems]
             };
         } catch (error) {
-            console.warn('Error saving viewer state:', error);
-        }
-    }
-
-    /**
-     * Get the current camera state for preservation
-     */
-    private getCurrentCameraState() {
-        if (!this.viewer || !this.viewer.camera) return null;
-
-        try {
-            const camera = this.viewer.camera;
-            const position = camera.positionCartographic;
-            
-            return {
-                longitude: position.longitude,
-                latitude: position.latitude,
-                height: position.height,
-                heading: camera.heading,
-                pitch: camera.pitch,
-                roll: camera.roll,
-                viewRectangle: camera.computeViewRectangle()
-            };
-        } catch (e) {
-            console.warn('Error getting camera state:', e);
-            return null;
+            console.error('Error saving viewer state:', error);
         }
     }
 
@@ -1193,96 +592,70 @@ export class ErdblickViewComponent implements AfterViewInit, OnDestroy {
      */
     private async destroyViewer(): Promise<void> {
         // Early return if viewer is already null or destroyed
-        if (!this.viewer || (typeof this.viewer.isDestroyed === 'function' && this.viewer.isDestroyed())) {
-            console.warn('Viewer already null or destroyed, skipping destruction');
+        if (this.viewStateService.isUnavailable() || this.viewStateService.isDestroyed()) {
+            console.debug('Viewer already null or destroyed, skipping destruction');
             return;
         }
-        
-        if (this._isDestroyingViewer) {
-            console.warn('Viewer already in destruction process.');
+
+        if (this.viewStateService.isDestroyingViewer) {
+            console.debug('Viewer already in destruction process.');
             return;
         }
-        this._isDestroyingViewer = true;
+        this.viewStateService.isDestroyingViewer = true;
 
         return new Promise((resolve) => {
             try {
                 // Clean up mouse handler first
                 if (this.mouseHandler) {
-                    try {
-                        if (!this.mouseHandler.isDestroyed()) {
-                            this.mouseHandler.destroy();
-                        }
-                    } catch (e) {
-                        console.warn('Error destroying mouse handler');
-                        console.debug(e);
+                    if (!this.mouseHandler.isDestroyed()) {
+                        this.mouseHandler.destroy();
                     }
                     this.mouseHandler = null;
                 }
-                
+
                 // Clean up collections and entities references
-                this.markerCollection = null;
+                this.markerService.markerCollection = null;
                 this.tileOutlineEntity = null;
                 this.openStreetMapLayer = null;
-                
+
                 // Clean the feature search visualization collection up
                 if (this.featureSearchService.visualization && !this.featureSearchService.visualization.isDestroyed()) {
-                    try {
-                        this.featureSearchService.visualization.destroy();
-                    } catch (e) {
-                        console.warn('Error destroying feature search visualization');
-                        console.debug(e);
-                    }
+                    this.featureSearchService.visualization.destroy();
                 }
-                
-                // Destroy viewer with multiple safety checks
-                if (this.viewer) {
-                    try {
-                        // Check if the viewer is still valid before operations
-                        if (typeof this.viewer.isDestroyed === 'function' && !this.viewer.isDestroyed()) {
-                            // Remove event listeners first
-                            if (this.viewer.camera) {
-                                try {
-                                    this.viewer.camera.changed.removeEventListener(this.cameraChangedHandler);
-                                    this.viewer.camera.moveStart.removeEventListener(this.cameraMoveStartHandler);
-                                    this.viewer.camera.moveEnd.removeEventListener(this.cameraMoveEndHandler);
-                                } catch (e) {
-                                    console.warn('Error removing camera event listeners');
-                                    console.debug(e);
-                                }
-                            }
 
-                            // Destroy the viewer - final safety check with try-catch
-                            if (typeof this.viewer.destroy === 'function' && !this.viewer.isDestroyed()) {
-                                try {
-                                    this.viewer.destroy();
-                                } catch (e) {
-                                    console.warn('Error calling viewer.destroy() - viewer may have been destroyed already:');
-                                    console.debug(e);
-                                }
-                            }
+                // Destroy viewer with multiple safety checks
+                if (this.viewStateService.viewer && this.viewStateService.isNotDestroyed()) {
+                        // Remove event listeners first
+                        if (this.viewStateService.viewer.camera) {
+                            this.viewStateService.viewer.camera.changed.removeEventListener(
+                                this.viewService.updateOnCameraChangedHandler
+                            );
+                            this.viewStateService.viewer.camera.moveStart.removeEventListener(
+                                this.cameraService.cameraMoveStartHandler
+                            );
+                            this.viewStateService.viewer.camera.moveEnd.removeEventListener(
+                                this.cameraService.cameraMoveEndHandler
+                            );
                         }
-                    } catch (e) {
-                        console.warn('Error during viewer destruction:');
-                        console.debug(e);
-                    }
-                    
+                        this.viewStateService.viewer.destroy();
+
                     // Clear viewer reference regardless
-                    this.viewer = null as any;
+                    this.viewStateService.viewer = null as any;
                 }
-                
+
                 // Small delay to ensure DOM cleanup completes
-            setTimeout(() => {
-                    this._isDestroyingViewer = false;
+                setTimeout(() => {
+                    this.viewStateService.isDestroyingViewer = false;
                     resolve();
                 }, 100);
-                
+
             } catch (error) {
                 console.error('Error during viewer destruction:', error);
-                this._isDestroyingViewer = false;
+                this.viewStateService.isDestroyingViewer = false;
                 // Clear references even on error
-                this.viewer = null as any;
+                this.viewStateService.viewer = null as any;
                 this.mouseHandler = null;
-                this.markerCollection = null;
+                this.markerService.markerCollection = null;
                 this.tileOutlineEntity = null;
                 this.openStreetMapLayer = null;
                 resolve(); // Continue anyway
@@ -1297,8 +670,8 @@ export class ErdblickViewComponent implements AfterViewInit, OnDestroy {
         return new Promise((resolve, reject) => {
             try {
                 const mapProjection = is2D ? new WebMercatorProjection() : new GeographicProjection();
-                
-                this.viewer = new Viewer("mapViewContainer", {
+
+                this.viewStateService.viewer = new Viewer("mapViewContainer", {
                     baseLayerPicker: false,
                     animation: false,
                     geocoder: false,
@@ -1320,6 +693,7 @@ export class ErdblickViewComponent implements AfterViewInit, OnDestroy {
                 setTimeout(async () => {
                     try {
                         // Setup all viewer components
+                        this.viewStateService.isViewerInit.next(true);
                         await this.setupViewerComponents();
                         resolve();
                     } catch (error) {
@@ -1327,7 +701,7 @@ export class ErdblickViewComponent implements AfterViewInit, OnDestroy {
                         reject(error);
                     }
                 }, 100);
-                
+
             } catch (error) {
                 console.error('Error creating viewer:', error);
                 reject(error);
@@ -1340,94 +714,74 @@ export class ErdblickViewComponent implements AfterViewInit, OnDestroy {
      */
     private async setupViewerComponents(): Promise<void> {
         try {
-            this.setupSceneMode(this.is2DMode);
+            this.setupSceneMode(this.viewStateService.is2DMode);
 
             // Recreate OpenStreetMap layer
-            this.openStreetMapLayer = this.viewer.imageryLayers.addImageryProvider(this.getOpenStreetMapLayerProvider());
-            
+            this.openStreetMapLayer = this.viewStateService.viewer.imageryLayers.addImageryProvider(
+                this.getOpenStreetMapLayerProvider()
+            );
+
             // Recreate mouse handler
-            this.mouseHandler = new ScreenSpaceEventHandler(this.viewer.scene.canvas);
+            this.mouseHandler = new ScreenSpaceEventHandler(this.viewStateService.viewer.scene.canvas);
             this.setupEventHandlers();
             this.setupCameraHandlers();
             this.setupWheelHandler();
 
             // Set globe appearance
-            this.viewer.scene.globe.baseColor = new Color(0.1, 0.1, 0.1, 1);
+            this.viewStateService.viewer.scene.globe.baseColor = new Color(0.1, 0.1, 0.1, 1);
 
             // Remove fullscreen button
-            if (this.viewer.fullscreenButton && !this.viewer.fullscreenButton.isDestroyed()) {
-                this.viewer.fullscreenButton.destroy();
+            if (this.viewStateService.viewer.fullscreenButton &&
+                !this.viewStateService.viewer.fullscreenButton.isDestroyed()) {
+                this.viewStateService.viewer.fullscreenButton.destroy();
             }
 
             // Recreate marker collection
-            this.markerCollection = new BillboardCollection({
-                scene: this.viewer.scene
+            this.markerService.markerCollection = new BillboardCollection({
+                scene: this.viewStateService.viewer.scene
             });
-            this.viewer.scene.primitives.add(this.markerCollection);
+            this.viewStateService.viewer.scene.primitives.add(this.markerService.markerCollection);
 
             // Recreate feature search visualization collection for new viewer
             this.featureSearchService.visualization = new BillboardCollection({
-                scene: this.viewer.scene
+                scene: this.viewStateService.viewer.scene
             });
-            this.viewer.scene.primitives.add(this.featureSearchService.visualization);
+            this.viewStateService.viewer.scene.primitives.add(this.featureSearchService.visualization);
 
             // Re-render existing search results if any
             if (this.featureSearchService.searchResults.length > 0) {
-                this.renderFeatureSearchResultTree(this.mapService.zoomLevel.getValue());
+                this.markerService.renderFeatureSearchResultTree(this.mapService.zoomLevel.getValue());
             }
-            
+
         } catch (error) {
             console.error('Error during viewer component initialization:', error);
             throw error;
         }
     }
 
-    /**
-     * Restore markers based on current parameters
-     * This handles the case where parameter subscriptions fired before markerCollection was ready
-     */
-    private restoreParameterMarker() {
-        try {
-            const currentParams = this.parameterService.parameters.getValue();
-            if (currentParams.marker && currentParams.markedPosition.length === 2) {
-                const markerPosition = Cartesian3.fromDegrees(
-                    Number(currentParams.markedPosition[0]),
-                    Number(currentParams.markedPosition[1])
-                );
-                const success = this.addMarker(markerPosition);
-                if (success) {
-                    console.debug('Parameter-driven focus marker restored after viewer reinitialization');
-            } else {
-                    console.warn('Failed to restore parameter-driven focus marker');
-                }
-            }
-        } catch (error) {
-            console.warn('Error restoring parameter markers:', error);
-        }
-    }
 
     /**
      * Setup camera event handlers
      */
     private setupCameraHandlers() {
-        this.viewer.camera.percentageChanged = 0.1;
-        this.viewer.camera.changed.addEventListener(this.cameraChangedHandler);
-        this.viewer.camera.moveStart.addEventListener(this.cameraMoveStartHandler);
-        this.viewer.camera.moveEnd.addEventListener(this.cameraMoveEndHandler);
+        this.viewStateService.viewer.camera.percentageChanged = 0.1;
+        this.viewStateService.viewer.camera.changed.addEventListener(this.viewService.updateOnCameraChangedHandler);
+        this.viewStateService.viewer.camera.moveStart.addEventListener(this.cameraService.cameraMoveStartHandler);
+        this.viewStateService.viewer.camera.moveEnd.addEventListener(this.cameraService.cameraMoveEndHandler);
     }
 
     /**
      * Setup custom wheel handler for 2D mode
      */
     private setupWheelHandler() {
-        this.viewer.scene.canvas.addEventListener('wheel', (event: WheelEvent) => {
-            if (this.is2DMode) {
+        this.viewStateService.viewer.scene.canvas.addEventListener('wheel', (event: WheelEvent) => {
+            if (this.viewStateService.is2DMode) {
                 event.preventDefault();
                 event.stopPropagation();
-                
+
                 const zoomFactor = event.deltaY > 0 ? 1.1 : 0.9;
-                
-                this.zoom2D(zoomFactor);
+
+                this.cameraService.zoom2D(zoomFactor);
             }
         });
     }
@@ -1436,460 +790,104 @@ export class ErdblickViewComponent implements AfterViewInit, OnDestroy {
      * Restore viewer state after reinitialization
      */
     private restoreViewerState() {
-        if (!this.viewerState || !this.viewer) {
-            console.warn('Cannot restore viewer state: missing state or viewer');
+        if (!this.viewStateService.viewerState || !this.viewStateService.viewer) {
+            console.debug('Cannot restore viewer state: missing state or viewer');
             return;
         }
 
         try {
             // Restore OpenStreetMap layer
             if (this.openStreetMapLayer) {
-                this.openStreetMapLayer.alpha = this.viewerState.openStreetMapLayerAlpha;
-                this.openStreetMapLayer.show = this.viewerState.openStreetMapLayerShow;
+                this.openStreetMapLayer.alpha = this.viewStateService.viewerState.openStreetMapLayerAlpha;
+                this.openStreetMapLayer.show = this.viewStateService.viewerState.openStreetMapLayerShow;
             }
 
             // Restore markers
-            if (this.markerCollection && this.viewerState.markerPositions.length > 0) {
-                this.viewerState.markerPositions.forEach(position => {
-                    this.addMarker(position);
+            if (this.markerService.markerCollection && this.viewStateService.viewerState.markerPositions.length > 0) {
+                this.viewStateService.viewerState.markerPositions.forEach(position => {
+                    this.markerService.addMarker(position);
                 });
             }
 
             // Restore camera state
-            this.restoreCameraState();
+            this.cameraService.restoreCameraState(this.viewStateService.viewerState);
 
             // Restore menu items
-            this.menuItems = this.viewerState.menuItems;
+            this.menuItems = this.viewStateService.viewerState.menuItems;
 
             // Clear saved state
-            this.viewerState = null;
-            
+            this.viewStateService.viewerState = null;
+
             // Force a render to ensure everything is displayed
-            if (this.viewer && this.viewer.scene) {
-        this.viewer.scene.requestRender();
+            if (this.viewStateService.viewer && this.viewStateService.viewer.scene) {
+                this.viewStateService.viewer.scene.requestRender();
             }
-            
+
         } catch (error) {
             console.error('Error restoring viewer state:', error);
             // Clear state on error to prevent future issues
-            this.viewerState = null;
-        }
-    }
-
-    /**
-     * Restore camera state from saved state
-     */
-    private restoreCameraState() {
-        if (!this.viewerState?.cameraState || !this.viewer || !this.viewer.camera) {
-            console.warn('Cannot restore camera state: missing state, viewer, or camera');
-            return;
-        }
-
-        try {
-            const cameraState = this.viewerState.cameraState;
-            this.ignoreNextCameraUpdate = true;
-
-            if (this.is2DMode) {
-                // For 2D mode, use view rectangle if available
-                if (cameraState.viewRectangle) {
-                    this.viewer.camera.setView({
-                        destination: cameraState.viewRectangle
-                    });
-                } else {
-                    // Fallback to center position
-                    this.viewer.camera.setView({
-                        destination: Rectangle.fromRadians(
-                            cameraState.longitude - 0.01,
-                            cameraState.latitude - 0.01,
-                            cameraState.longitude + 0.01,
-                            cameraState.latitude + 0.01
-                        )
-                    });
-                }
-            } else {
-                // For 3D mode, restore full camera state
-                this.viewer.camera.setView({
-                    destination: Cartesian3.fromRadians(
-                        cameraState.longitude, 
-                        cameraState.latitude, 
-                        cameraState.height
-                    ),
-                    orientation: {
-                        heading: cameraState.heading,
-                        pitch: cameraState.pitch,
-                        roll: cameraState.roll
-                    }
-                });
-            }
-        } catch (error) {
-            console.error('Error restoring camera state:', error);
+            this.viewStateService.viewerState = null;
         }
     }
 
     private setupSceneMode(is2D: boolean) {
-        this.is2DMode = is2D;
-        if (this.is2DMode) {
-            this.viewer.scene.mode = SceneMode.SCENE2D;
-            this.setup2DConstraints();
+        this.viewStateService.is2DMode = is2D;
+        if (this.viewStateService.is2DMode) {
+            this.viewStateService.viewer.scene.mode = SceneMode.SCENE2D;
+            this.viewService.setup2DModeConstraints();
         } else {
-            this.viewer.scene.mode = SceneMode.SCENE3D;
-            this.setup3DConstraints();
+            this.viewStateService.viewer.scene.mode = SceneMode.SCENE3D;
+            this.viewService.setup3DModeConstraints();
         }
     }
-
-    private setup2DConstraints() {
-        // Enable 2D map interactions
-        const scene = this.viewer.scene;
-        
-        // Disable camera rotation and tilting in 2D
-        scene.screenSpaceCameraController.enableRotate = false;
-        scene.screenSpaceCameraController.enableTilt = false;
-        
-        // Enable standard 2D interactions
-        scene.screenSpaceCameraController.enableTranslate = true;
-        scene.screenSpaceCameraController.enableZoom = false; // Disable Cesium's zoom to use only our custom handler
-        scene.screenSpaceCameraController.enableLook = false;
-        
-        // Set zoom constraints for 2D mode (not used since we disabled zoom)
-        scene.screenSpaceCameraController.minimumZoomDistance = 100;
-        scene.screenSpaceCameraController.maximumZoomDistance = 50000000;
-    }
-
-    private setup3DConstraints() {
-        // Re-enable full 3D camera controls
-        const scene = this.viewer.scene;
-        
-        scene.screenSpaceCameraController.enableRotate = true;
-        scene.screenSpaceCameraController.enableTilt = true;
-        scene.screenSpaceCameraController.enableTranslate = true;
-        scene.screenSpaceCameraController.enableZoom = true; // Re-enable Cesium's zoom for 3D mode
-        scene.screenSpaceCameraController.enableLook = true;
-        
-        // Reset zoom constraints for 3D mode
-        scene.screenSpaceCameraController.minimumZoomDistance = 1;
-        scene.screenSpaceCameraController.maximumZoomDistance = 50000000;
-    }
-
-    /**
-     * Calculate the minimum view rectangle height for 2D mode (1 meter in degrees)
-     */
-    private getMinViewRectangleHeight(): number {
-        // The Minimum corresponds to about 100 meters altitude using natural scaling
-        // This provides a reasonable minimum zoom level
-        const earthRadius = this.viewer.scene.globe.ellipsoid.maximumRadius; // Use viewer's ellipsoid maximum radius
-        const minAltitude = 100; // meters
-        return 2 * Math.atan(minAltitude / (2 * earthRadius));
-    }
-
-    /**
-     * Calculate the maximum view rectangle height for 2D mode (45 degrees)
-     */
-    private getMaxViewRectangleHeight(): number {
-        // Limit to 45 degrees to prevent excessive zoom out
-        // This provides a reasonable world view without black bars
-        return CesiumMath.toRadians(45);
-    }
-
-    /**
-     * Get movement distance for 2D mode based on current viewport
-     */
-    private get2DMovementDistance(): {longitudeOffset: number, latitudeOffset: number} {
-        const currentView = this.viewer.camera.computeViewRectangle();
-        if (!currentView) {
-            // Fallback to default movement if view can't be computed
-            return {
-                longitudeOffset: this.parameterService.cameraMoveUnits,
-                latitudeOffset: this.parameterService.cameraMoveUnits
-            };
-        }
-
-        const currentWidth = currentView.east - currentView.west;
-        const currentHeight = currentView.north - currentView.south;
-        
-        // Move by 10% of the current view size (adjust this percentage as needed)
-        const movementPercentage = 0.1;
-        
-        // Calculate movement distances
-        const longitudeOffset = CesiumMath.toDegrees(currentWidth * movementPercentage);
-        const latitudeOffset = CesiumMath.toDegrees(currentHeight * movementPercentage);
-        
-        // Clamp movements to reasonable values
-        const maxLonMovement = 45; // Maximum 45 degrees longitude movement
-        const minLonMovement = 0.001;
-        const maxLatMovement = 45; // Maximum 45 degrees latitude movement
-        const minLatMovement = 0.001;
-        
-        return {
-            longitudeOffset: Math.max(minLonMovement, Math.min(maxLonMovement, longitudeOffset)),
-            latitudeOffset: Math.max(minLatMovement, Math.min(maxLatMovement, latitudeOffset))
-        };
-    }
-
-    /**
-     * 2D zoom using height-based approach with proper limit conversion
-     */
-    private zoom2D(zoomFactor: number): void {
-        if (!this.viewer || !this.viewer.camera) {
-            console.warn('Cannot zoom in 2D: viewer not available');
-            return;
-        }
-
-        try {
-            // Check if viewer is destroyed
-            if (typeof this.viewer.isDestroyed === 'function' && this.viewer.isDestroyed()) {
-                console.warn('Cannot zoom in 2D: viewer is destroyed');
-                return;
-            }
-
-        const camera = this.viewer.camera;
-        
-        // Get current camera height
-        const currentHeight = camera.positionCartographic.height;
-        const newHeight = currentHeight * zoomFactor;
-        
-        // Convert height to equivalent view rectangle height using natural relationship
-        const earthRadius = this.viewer.scene.globe.ellipsoid.maximumRadius; // Use viewer's ellipsoid maximum radius
-        const newViewRectHeight = 2 * Math.atan(newHeight / (2 * earthRadius));
-        
-        // Apply zoom limits based on view rectangle height
-        const minViewRectHeight = this.getMinViewRectangleHeight();
-        const maxViewRectHeight = this.getMaxViewRectangleHeight();
-        
-        let clampedViewRectHeight = newViewRectHeight;
-        let wasClampedToLimit = false;
-        
-        if (newViewRectHeight < minViewRectHeight) {
-            clampedViewRectHeight = minViewRectHeight;
-            wasClampedToLimit = true;
-        } else if (newViewRectHeight > maxViewRectHeight) {
-            clampedViewRectHeight = maxViewRectHeight;
-            wasClampedToLimit = true;
-        }
-        
-        // Convert the clamped view rectangle height back to camera height
-        const clampedHeight = (2 * earthRadius) * Math.tan(clampedViewRectHeight / 2);
-        
-        if (wasClampedToLimit) {
-            console.debug('Zoom clamped to view rectangle limits');
-        }
-        
-        // Set new camera height directly while preserving position
-        const currentPos = camera.positionCartographic;
-        const newPosition = Cartesian3.fromRadians(
-            currentPos.longitude,
-            currentPos.latitude,
-            clampedHeight
-        );
-        
-        // Ignore the camera change event for this programmatic zoom
-        this.ignoreNextCameraUpdate = true;
-        
-        camera.setView({
-            destination: newPosition
-        });
-        } catch (error) {
-            console.warn('Error in 2D zoom:', error);
-        }
-    }
-
-    // Store event handler references for proper cleanup
-    private cameraChangedHandler = () => {
-        try {
-            // Check if viewer is still valid
-            if (!this.viewer || !this.viewer.camera) {
-                console.error('cameraChangedHandler: viewer or camera is null');
-                return;
-            }
-            
-            if (typeof this.viewer.isDestroyed === 'function' && this.viewer.isDestroyed()) {
-                console.error('cameraChangedHandler: viewer is destroyed');
-                return;
-            }
-            
-            // Check for extreme coordinates that might cause issues
-            const cameraPos = this.viewer.camera.positionCartographic;
-            const lon = CesiumMath.toDegrees(cameraPos.longitude);
-            const lat = CesiumMath.toDegrees(cameraPos.latitude);
-            if (!isFinite(lon) || !isFinite(lat) || Math.abs(lon) > 180 || Math.abs(lat) > 90) {
-                console.error('Invalid camera coordinates detected:', { lon, lat });
-                return;
-            }
-            
-            if (!this.ignoreNextCameraUpdate) {
-                if (this.is2DMode) {
-                    this.parameterService.set2DCameraState(this.viewer.camera);
-                } else {
-                    this.parameterService.setCameraState(this.viewer.camera);
-                }
-            }
-            
-            this.ignoreNextCameraUpdate = false;
-            this.updateViewport();
-            
-        } catch (error) {
-            console.error('Error in cameraChangedHandler:', error);
-        }
-    };
-
-    private cameraMoveStartHandler = () => {
-        this.cameraIsMoving = true;
-    };
-
-    private cameraMoveEndHandler = () => {
-        this.cameraIsMoving = false;
-    };
 
     /**
      * Component cleanup when destroyed
      */
     ngOnDestroy() {
-        console.log('ErdblickViewComponent: cleaning up resources');
-        
+        console.debug('ErdblickViewComponent: cleaning up resources');
+
         // Don't allow mode changes during destruction
-        this._isChangingMode = true;
-        
+        this.viewStateService.isChangingMode = true;
+
         // Clean up resources without async to avoid hanging
-        if (this.mouseHandler) {
-            try {
+        try {
+            if (this.mouseHandler) {
                 if (!this.mouseHandler.isDestroyed()) {
                     this.mouseHandler.destroy();
                 }
-            } catch (e) {
-                console.warn('Error destroying mouse handler:', e);
+                this.mouseHandler = null;
             }
-            this.mouseHandler = null;
-        }
-        
-        if (this.viewer) {
-            try {
-                if (typeof this.viewer.isDestroyed === 'function' && !this.viewer.isDestroyed()) {
-                    // Remove event listeners before destroying
-                    if (this.viewer.camera) {
-                        try {
-                            this.viewer.camera.changed.removeEventListener(this.cameraChangedHandler);
-                            this.viewer.camera.moveStart.removeEventListener(this.cameraMoveStartHandler);
-                            this.viewer.camera.moveEnd.removeEventListener(this.cameraMoveEndHandler);
-                        } catch (e) {
-                            console.warn('Error removing camera event listeners in ngOnDestroy:', e);
-                        }
-                    }
-                    
-                    this.viewer.destroy();
+
+            if (this.viewStateService.viewer && this.viewStateService.isNotDestroyed()) {
+                // Remove event listeners before destroying
+                if (this.viewStateService.viewer.camera) {
+                    this.viewStateService.viewer.camera.changed.removeEventListener(
+                        this.viewService.updateOnCameraChangedHandler
+                    );
+                    this.viewStateService.viewer.camera.moveStart.removeEventListener(
+                        this.cameraService.cameraMoveStartHandler
+                    );
+                    this.viewStateService.viewer.camera.moveEnd.removeEventListener(
+                        this.cameraService.cameraMoveEndHandler
+                    );
                 }
-            } catch (e) {
-                console.warn('Error destroying viewer in ngOnDestroy:', e);
+                this.viewStateService.viewer.destroy();
+                this.viewStateService.viewer = null as any;
             }
-            this.viewer = null as any;
+        } catch (e) {
+            console.error('Error in ngOnDestroy:', e);
         }
-        
+
         // Clear all references
         this.mouseHandler = null;
         this.openStreetMapLayer = null;
-        this.markerCollection = null;
+        this.markerService.markerCollection = null;
         this.tileOutlineEntity = null;
-        this.viewerState = null;
-        
+        this.viewStateService.viewerState = null;
+
         // Reset flags
-        this._isChangingMode = false;
-        this._isDestroyingViewer = false;
-    }
-
-    /**
-     * Compute viewport rectangle using robust sampling along edges
-     * Workaround for Cesium's computeViewRectangle failing in 2D WebMercator
-     */
-    private computeRobustViewRectangle(canvas: HTMLCanvasElement): Rectangle | undefined {
-        try {
-            const samplePoints: Array<{x: number, y: number}> = [];
-            const width = canvas.clientWidth;
-            const height = canvas.clientHeight;
-            const steps = 10; // Sample every 10th pixel along edges
-            
-            // Sample along edges and add interior points
-            for (let x = 0; x <= width; x += Math.max(1, Math.floor(width / steps))) {
-                samplePoints.push({x, y: 0});      // top edge
-                samplePoints.push({x, y: height}); // bottom edge
-            }
-            for (let y = 0; y <= height; y += Math.max(1, Math.floor(height / steps))) {
-                samplePoints.push({x: 0, y});      // left edge
-                samplePoints.push({x: width, y});  // right edge
-            }
-            
-            // Add interior points for better coverage
-            samplePoints.push({x: width/2, y: height/2});   // center
-            samplePoints.push({x: width/4, y: height/4});   // quarter points
-            samplePoints.push({x: 3*width/4, y: height/4});
-            samplePoints.push({x: width/4, y: 3*height/4});
-            samplePoints.push({x: 3*width/4, y: 3*height/4});
-            
-            const validCoordinates: Array<{lon: number, lat: number}> = [];
-            
-            // Try to pick ellipsoid intersection for each sample point
-            for (const point of samplePoints) {
-                try {
-                    const cartesian = this.viewer.camera.pickEllipsoid(
-                        new Cartesian2(point.x, point.y), 
-                        this.viewer.scene.globe.ellipsoid
-                    );
-                    
-                    if (cartesian) {
-                        const cartographic = Cartographic.fromCartesian(cartesian);
-                        const lon = CesiumMath.toDegrees(cartographic.longitude);
-                        const lat = CesiumMath.toDegrees(cartographic.latitude);
-                        
-                        // Validate coordinates
-                        if (isFinite(lon) && isFinite(lat) && 
-                            lon >= -180 && lon <= 180 && 
-                            lat >= -90 && lat <= 90) {
-                            validCoordinates.push({lon, lat});
-                        }
-                    }
-                } catch (error) {
-                    // Skip invalid points
-                }
-            }
-            
-            if (validCoordinates.length < 3) {
-                return undefined;
-            }
-            
-            // Find bounding box of valid coordinates
-            const lons = validCoordinates.map(coord => coord.lon);
-            const lats = validCoordinates.map(coord => coord.lat);
-            
-            let minLon = Math.min(...lons);
-            let maxLon = Math.max(...lons);
-            let minLat = Math.min(...lats);
-            let maxLat = Math.max(...lats);
-            
-            // Handle longitude wrapping around ±180°
-            if (maxLon - minLon > 180) {
-                // Likely crossing antimeridian, use camera position as reference
-                const cameraPos = this.viewer.camera.positionCartographic;
-                const cameraLon = CesiumMath.toDegrees(cameraPos.longitude);
-
-                const areLonsCorrect = validCoordinates.every(coord => Math.abs(coord.lon - cameraLon) <= 180);
-
-                if (!areLonsCorrect) {
-                    const reprojectedLons = validCoordinates.map(coord => {
-                        let lon = coord.lon;
-                        if (lon - cameraLon > 180) {
-                            lon -= 360;
-                        } else if (cameraLon - lon > 180) {
-                            lon += 360;
-                        }
-                        return lon;
-                    });
-                    minLon = Math.min(...reprojectedLons);
-                    maxLon = Math.max(...reprojectedLons);
-                }
-            }
-            
-            return Rectangle.fromDegrees(minLon, minLat, maxLon, maxLat);
-            
-        } catch (error) {
-            console.error('Error in robust rectangle calculation:', error);
-            return undefined;
-        }
+        this.viewStateService.isChangingMode = false;
+        this.viewStateService.isDestroyingViewer = false;
     }
 }
